+++ src/main.cpp
@@ -1,5 +1,6 @@
 // Entry Point for the SAMFILE Parser
 
+#include "sections.hpp"
 #include "settings.hpp"
 #include "version.hpp"
 #include <batch2.hpp>
@@ -11,6 +12,8 @@
 #include <sfp/runner.hpp>
 #include <strutil.hpp>
 #include <unordered_map>
+#include "test.hpp"
+
 
 using MAP = std::unordered_map<std::string, std::string>;
 using KVPMAP = KeyValueParser2::KeyValueStore<MAP>;
@@ -25,6 +28,8 @@ enum class SECTION : std::uint8_t
     DATA_JSON
 };
 
+
+// Function to print the small help message when no enough arguments
 int help()
 {
     std::cout << "Usage: sfp [task]\n"
@@ -32,6 +37,8 @@ int help()
     return 0;
 }
 
+
+// Function to print the full help messages
 int fullhelp()
 {
     std::cout << "Help for sfp Version " BUILD_VERSION " at " BUILD_DATE
@@ -39,6 +46,8 @@ int fullhelp()
     return 0;
 }
 
+
+// Function to load the samfile
 std::string loadsamfile(const std::string &filename)
 {
     std::ifstream file(filename);
@@ -52,6 +61,8 @@ std::string loadsamfile(const std::string &filename)
     return buffer.str();
 }
 
+
+// Function to run a samfile
 int run_samfile2(const std::string &content, const Config &conf,
                  const std::string &cmd)
 {
@@ -72,194 +83,21 @@ int run_samfile2(const std::string &content, const Config &conf,
     return run_task(tasks, cmd, visited, state, conf);
 }
 
+
+// Main function to start the program
 int main(int argc, char *argv[])
 {
-#pragma region DEBUG TESTS
 
-#ifndef NNDEBUG
-    // RUN Tests
-
-    std::string test_samfile_content = R"(
-#define VERSION 2
-
-#ifdef SECTION_SAMFILE
-
-# Build and run macro parser
-mp:
-    CD libs/macroparser
-    RUN cmake -S . -B build
-    RUN ln -sf build/compile_commands.json compile_commands.json
-    RUN cmake --build build
-    RUN ./build/mylib_test
-
-# Build the engine with CMake
-build:
-    RUN cmake -S . -B build
-    RUN ln -sf build/compile_commands.json compile_commands.json
-    RUN cmake --build build
-
-b:
-    TASK build
-
-be:
-    RUN cmake --build build
-    RUN ./build/mygame
-
-# Task to compile the Go DLL (not used yet)
-buildgo:
-    CD libs/libgodll
-    ENV CC="zig cc"
-    ENV CXX="zig c++"
-    SHELL go build -buildmode=c-shared -o libgo.dll main.go
-
-run:
-    TASK build
-    RUN ./build/mygame
-
-r:
-    TASK run
-
-buildcrablang:
-    CD libs/libcrablang
-    RUN cargo build --release
-    CD ../..
-    CP libs/libcrablang/target/release/libcrablang.dll out/build/x64-Debug/libcrablang.dll
-    CP libs/libcrablang/target/release/libcrablang.dll out/build/x64-Release/libcrablang.dll
-    CP libs/libcrablang/target/release/libcrablang.dll export/libcrablang.dll
-
-# Task for the export
-export:
-    TASKWIN exportwin
-    TASKLIN exportlin
-
-exportlin:
-    ECHO "Running Export for Linux"
-    RM export
-    MKDIR export
-    #TASK buildcrablang
-    TASK helpsite
-    TASK buildsinglepages
-    CP resources export/resources
-    CP build/mygame export/mygame
-    WRITE export/mygame_climenu "#!/bin/bash\n\necho \"Running in Cli Selection Mode:\"\n\nmygame --no-ui --select-menu\n\npause\n"
-    WRITE export/mygame_cli "#!/bin/bash\n\necho \"Running in Cli Mode:\"\n\nmygame --no-ui \"$@\"\n"
-    WRITE export/fling "#!/bin/bash\n\necho \"Running FLING SOURCE:\"\necho.\n\nmygame --no-ui fling \"$@\"\\n"
-    WRITE export/lua "#!/bin/bash\n\necho \"Running LUA SOURCE:\"\necho.\n\nmygame --no-ui lua \"$@\"\\n"
-    WRITE export/lua "#!/bin/bash\n\nmygame --no-ui subm \"$@\"\\n"
-
-# Task for the Windows Export
-exportwin:
-    ECHO "Running Export for Windows"
-    RM export
-    MKDIR export
-    #TASK buildcrablang
-    TASK helpsite
-    TASK buildsinglepages
-    CP resources export/resources
-    CP out/build/x64-Release/mygame.exe export/mygame.exe
-    WRITE export/open_a_console_here.bat "@cmd\n"
-    WRITE export/mygame_climenu.bat "@echo off\n\necho Running in Cli Selection Mode:\n\ncall mygame --no-ui --select-menu\n\npause\n"
-    WRITE export/mygame_cli.bat "@echo off\n\necho Running in Cli Mode:\n\ncall mygame --no-ui %*\n"
-    WRITE export/fling.bat "@echo off\n\necho Running FLING SOURCE:\necho.\n\ncall mygame --no-ui fling %*\n"
-    WRITE export/lua.bat "@echo off\n\necho Running LUA SOURCE:\necho.\n\ncall mygame --no-ui lua %*\n"
-    WRITE export/git-subm.bat "@echo off\n\ncall mygame --no-ui subm %*\n"
-
-e:
-    TASK export
-
-# Init Task
-init:
-    RUN l2 clonesubm
-    RM resources/editor/htmlpages/.git
-    RM resources/editor/htmlpages/.gitattributes
-    RUN bun i
-
-i:
-    TASK init
-
-# to build to static HTML Page
-helpsite:
-    RUN l2 view
-    MV .samengine/links.md libs/docs/pages/links.md
-    CD libs/docs
-    RUN bun sam-cli minisite
-    MV index.html ../../resources/editor/notice.html
-    CD ../../
-
-# Build Single Pages
-buildsinglepages:
-	ECHO RUN THIS COMMAND ONLY BEFORE RELEASES !!!
-    CD libs
-    CD vue
-    RUN bun i
-    RUN bun run build
-    CD ..
-    CD react
-    RUN bun i
-    RUN bun run build
-    CD ..
-    CD svelte
-    RUN bun i
-    RUN bun run build
-    CD ..
-    CD ..
-    MV libs/vue/dist/index.html resources/editor/htmlpages/vue.html
-    MV libs/react/dist/index.html resources/editor/htmlpages/react.html
-    MV libs/svelte/dist/index.html resources/editor/htmlpages/svelte.html
-
-bs:
-    TASK buildsinglepages
-
-bl:
-    TASK buildcrablang
-
-// Done
+#ifdef NNDEBUG
+	test();
 #endif
 
-
-#ifdef SECTION_FLING
-
-// This is fling Code
-print("Running FLING SOURCE:")
-
-#endif
-
-
-#ifdef SECTION_BATCH2
-echo Hallo
-echo DOne
-#endif
-
-
-#ifdef SECTION_DATA
-
-#endif
-
-
-#ifdef SECTION_SETTINGS
-# Done
-#endif
-
-
-)";
-
-    MacroParser parser2;
-    parser2.add_macro("SECTION_FLING", {});
-
-    std::cout << parser2.parse_macros(test_samfile_content) << std::endl;
-
-    parser2.remove_macro("SECTION_FLING");
-    parser2.add_macro("SECTION_SAMFILE", {});
-
-    std::cout << parser2.parse_macros(test_samfile_content) << std::endl;
-#endif
-
-#pragma endregion
-
+    // Check the first argument
     std::string arg1 = (argc > 1) ? argv[1] : "";
 
     // Check for bat2 to execute them
     // When the first arg ends with .bat2
+    //
     if (strutil::ends_with(arg1, ".bat2"))
     {
         std::ifstream file(arg1);
@@ -278,6 +116,7 @@ echo DOne
 
     // Check for macroparsing
     // pm -> parse macros
+    //
     if (arg1 == "-pm")
     {
         if (argc >= 4)
@@ -314,6 +153,7 @@ echo DOne
     }
 
     // Print version
+    //
     if (arg1 == "--version" || arg1 == "-v")
     {
         std::cout << BUILD_VERSION "\n";
@@ -321,13 +161,18 @@ echo DOne
     }
 
     // Help
+    //
     if (arg1 == "--help" || arg1 == "-h")
     {
         return fullhelp();
     }
 
+    // Name for the loaded samfile
     std::string filename = "samfile";
 
+    // Check if the first arg end with .samfile because then a samfile will
+    // be loaded instead of a task
+    //
     if (strutil::ends_with(arg1, ".samfile"))
     {
         filename = arg1;
@@ -343,13 +188,19 @@ echo DOne
     int version = 0;
 
     // Check if Version exists
+    //
     if (parser.contains_macro("VERSION"))
     {
         auto macro = parser.get_macro("VERSION");
         version = std::stoi(macro.body);
     }
 
+    // Parse Sections
+    //
+    auto sections = parse_Sections(content);
+
     // OLD SAMFILE INTERPRETER
+    //
     if (version == 0)
     {
         if (argc < 2)
@@ -358,37 +209,32 @@ echo DOne
             return help();
         }
 
-        parser.clear_macros();
-
         // Run the old interpreter
         return run_samfile2(
-            parser.parse_macros(buildin_samfile_content + "\n\n" + content), {},
+            buildin_samfile_content + "\n\n" + sections[""], {},
             argv[1]);
     }
 
     // Samfile Version 2 Interpreter
+    //
     if (version == 2)
     {
         if (argc < 2)
         {
-            parser.add_macro("SECTION_BATCH2", {});
-            std::string input = parser.parse_macros(content);
-            auto tokens = batch2::tokenize(input);
+            auto tokens = batch2::tokenize(sections["BATCH2"]);
             batch2::Interpreter2 interp;
             return interp.execute(tokens);
         }
 
         KVPMAP seting;
-        parser.add_macro("SECTION_SETTINGS", {});
-        std::string set = parser.parse_macros(content);
-        seting = KeyValueParser2::parse_kvp2(set);
-        parser.remove_macro("SECTION_SETTINGS");
-        parser.add_macro("SECTION_SAMFILE", {});
+        seting = KeyValueParser2::parse_kvp2(sections["SETTINGS"]);
         return run_samfile2(
-            parser.parse_macros(buildin_samfile_content + "\n\n" + content), {},
+            buildin_samfile_content + "\n\n" + sections["SAMFILES"], {},
             argv[1]);
     }
 
+    // An error appeared when the program got until here
+    //
     std::cerr << RED << "Wrong samfile Version selected: " << version << END
               << "\n";
     return 1;